refactor: move codes associated with query parsing into `QueryTree`
